Moved ServiceClient to its own set of files
The service.hpp and service_impl.hpp files were getting very long and
unwieldy due to the large number of different classes they contained.
The ServiceClient class in particular is the longest, and most often
needs to be consulted by users since it defines the API clients interact
with, so I moved it to its own set of header/impl files.

Also, in a minor fix, removed a reference to the nonexistant
service_server_api.hpp from the documentation comments at the top of
service.hpp. This seems to refer to a much older design for Casacade
(where there was something called an 'onData()' function on servers) and
is confusing for users.
